requirements.txt
streamlit
pdfplumber
pytesseract
requests
pinecone-client
Pillow
python-docx
